keyboard, display, interrupts, and debugging (#12)
This commit includes a lot of debugging changes and some structural
changes. I added in a FFI to the reference implementation at
emulators101 so that we can view the run of both our state and the
reference implementation during run (and halt on the
differences). Made some changes to enable the functionality of the
display and the keyboard inputs which now work allowing the play of a
space invaders game. The game play is a little choppy, changing around
some of the values in lib.rs to either increase the number of updates
or increase the number of yields to the main javascript thread could
help there.